close window handler added
Added do nothing on close to frame, and also action listener to show popUp message  when X button in upper right comer is clicked. Previously then window close this way program was not terminated and hanging in memory.... with time a lot of processes  accumulated. This may cause clicking glitches at table headers to sort columns.