comment out DropDownItems of vbankNames
new bank added to comboBoxlist,
widow for new bank need to be created.
DropDownItems will be moved to dropboxes initializing class of file or interface.
All work as it was with DropDownItems object created in here. 
This object will be move to logic class, or creator  class, or can be used in constructor.